[se cambia el tipo de dato a decimal para guardar dinero]

+++ SunatIntegration.Application/DTOs/Sunat/ExchangeRateDto.cs
@@ -4,7 +4,7 @@ namespace SunatIntegration.Application.DTOs.Sunat
     public class ExchangeRateDto
     {
         public DateTime? DatePublic { get; set; }
-        public double? PriceSales { get; set; }
-        public double? Pricepurchase { get; set; }
+        public decimal? PriceSales { get; set; }
+        public decimal? Pricepurchase { get; set; }
     }
 }
