+++ SunatIntegration.Domain/Entities/SunatExchangeRate.cs
@@ -5,8 +5,8 @@ namespace SunatIntegration.Domain.Entities
     {
         public int Id { get; set; }
         public DateTime DatePublic { get; set; }
-        public double? PriceSales { get; set; }
-        public double? Pricepurchase { get; set; }
+        public decimal? PriceSales { get; set; }
+        public decimal? Pricepurchase { get; set; }
         public DateTime CreatedDate { get; set; }= DateTime.Now;
         public DateTime ModifiedDate { get; set; }= DateTime.Now;
